pyproject.toml
[project]
name = "samhall_hackathon"
version = "0.1.0"
description = "Solution for workers scheduling"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.122.0",
    "ortools>=9.14.6206",
    "pydantic-ai>=1.24.0",
    "python-magic>=0.4.27",
    "supabase>=2.24.0",
    "uuid>=1.30",
]
